Patch by Ulrich Arnold: Fixed a compiler warning on VC2008
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1308 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfFontMetricsObject.cpp
@@ -39,8 +39,8 @@ PdfFontMetricsObject::PdfFontMetricsObject( PdfObject* pFont, PdfObject* pDescri
     m_sName        = pDescriptor->GetDictionary().GetKey( "FontName" )->GetName();
     m_bbox         = pDescriptor->GetDictionary().GetKey( "FontBBox" )->GetArray();
     // OC 15.08.2010 BugFix: /FirstChar /LastChar /Widths are in the Font dictionary and not in the FontDescriptor
-    m_nFirst       = pFont->GetDictionary().GetKeyAsLong( "FirstChar", 0L );
-    m_nLast        = pFont->GetDictionary().GetKeyAsLong( "LastChar", 0L );
+    m_nFirst       = static_cast<int>(pFont->GetDictionary().GetKeyAsLong( "FirstChar", 0L ));
+    m_nLast        = static_cast<int>(pFont->GetDictionary().GetKeyAsLong( "LastChar", 0L ));
 	 // OC 15.08.2010 BugFix: GetIndirectKey() instead of GetDictionary().GetKey() and "Widths" instead of "Width"
     m_width        = pFont->GetIndirectKey( "Widths" )->GetArray();
